examples/openai_agents_sdk/youtube_summary_agent/requirements.txt
python-dotenv
openai
google_api_python_client
youtube_transcript_api
openai_agents
ragaai_catalyst